fix some types
